meta: use QT_VERSION to mask bbappends
meta-qt6 provides QT_VERSION variable that can be used to identify
and then mask bbappends that are not suitable for older Qt versions.
This makes it possible to easily switch the meta-qt6 layer to a
different version and still keep all the bbappends in the layer.

Change-Id: I04091752cf5388e19afd4f78ec6e132155563ba9
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>
(cherry picked from commit 2a6f3786391e001a47b458ed5737deea167a027e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>