meta-tegra: update layer
Update to latest revision in dunfell branch.

There are two qtbase patches added in the updated OE4T/dunfell revision,
of which patch#1 is present in qtbase and patch#2 needs a rebase
to deal with patch fuzz. Both patches from meta-tegra bbmasked and
a rebased patch#2 added to tegra dynamic layers section.

Change-Id: I529e8ab9b53677e450ef2297f7b33c867ae19b71
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>